% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcode_validate.R
\name{postcode_validate}
\alias{postcode_validate}
\title{Validate a postcode/huisnummer with the Postcode API}
\usage{
postcode_validate(postcode, huisnummer, key = getOption("postcode_api_key"))
}
\arguments{
\item{postcode}{E.g. '1234AB'}

\item{huisnummer}{Integer}
}
\value{
A list with address, location information, or NULL if postcode/huisnummer not found.
}
\description{
Validate a postcode/huisnummer with the Postcode API
}
\details{
Set the API key with `options(postcode_api_key = "abcdef")` (see 1Password).
}
\examples{
\dontrun{
postcode_validate("5617BD", 42)
}
}
