% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postcodeValidateUI.R
\name{postcodeValidateUI}
\alias{postcodeValidateUI}
\alias{postcodeValidate}
\title{Shiny module for address validation with Postcode API}
\usage{
postcodeValidateUI(id, extra_fields = TRUE)

postcodeValidate(input, output, session)
}
\arguments{
\item{id}{Shiny input id}

\item{extra_fields}{If TRUE, adds extra input fields 'huisletter' and 'huisnummertoevoeging',
of which neither is used in address validation but will be added to the output.}

\item{input}{Shiny parameter, do not set.}

\item{output}{Shiny parameter, do not set.}

\item{session}{Shiny parameter, do not set.}
}
\description{
UI and Server module functions for PostCode API validation in Shiny applications.
}
\details{
See \code{\link{postcode_validate}} on how to set the API key.
}
\examples{
\dontrun{
  library(shiny)

ui <- fluidPage(
  useShinyjs(),
  postcodeValidateUI("test"),
  tags$hr(),
  verbatimTextOutput("txt_out")

)

server <- function(input, output, session) {
  out <- callModule(postcodeValidate, "test")

  output$txt_out <- renderPrint({
    out()
  })
}

shinyApp(ui, server)
}
}
